[Refactor] Rename  `_gantt_chart` module to `_plot_gantt_chart`
