+++ job_shop_lib/visualization/_gantt_chart_video_and_gif_creation.py
@@ -23,7 +23,7 @@ from job_shop_lib.dispatching import (
     HistoryObserver,
 )
 from job_shop_lib.dispatching.rules import DispatchingRuleSolver
-from job_shop_lib.visualization._gantt_chart import plot_gantt_chart
+from job_shop_lib.visualization._plot_gantt_chart import plot_gantt_chart
 
 
 PlotFunction = Callable[
