+++ job_shop_lib/visualization/_plot_gantt_chart.py
